Updated repos role to be much more complex
The role will remove a whitelist of repository names that should not belong. It will also determine if the base centos-release package is missing any files. If it is, it will then remove the package, run a script to determine where the latest package is located, download it, and install the centos-release package